Break perf energy models into multiple files.
